requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]

llama-index
llama-index-llms-gemini
llama-index-embeddings-gemini
google-generativeai

PyMuPDF
python-docx